backend_one/requirements.txt
python-dotenv
flask
pydantic
flask_sqlalchemy
sqlalchemy